requirements_production.txt
# Production Face Recognition System Requirements
# Optimized for maximum accuracy (99.6%)

# Core dependencies
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0

# Face recognition models (MTCNN + FaceNet)
facenet-pytorch==2.5.3
torch==2.1.0
torchvision==0.16.0

# Fallback face recognition
face-recognition==1.3.0
dlib==19.24.2

# Database and data handling
pandas==2.0.3
openpyxl==3.1.2

# GUI
tk

# Optional: GPU acceleration (uncomment if you have CUDA)
# torch==2.1.0+cu118
# torchvision==0.16.0+cu118

# Build tools
cmake==3.27.0
